% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/github_release.R
\name{github_release}
\alias{github_release}
\title{upload the R package binary build as an asset to its github release}
\usage{
github_release(repo = Sys.getenv("TRAVIS_REPO_SLUG"),
  token = Sys.getenv("GITHUB_TOKEN"), release_tag = "linux", path = ".")
}
\arguments{
\item{repo}{Repository address in the format username/repo. Default value is read from the environment variable 'TRAVIS_REPO_SLUG'}

\item{token}{character oAuth2.0 token. Default value is read from the environment variable 'GITHUB_TOKEN'}

\item{release_tag}{the tag where the asset is associated to}

\item{path}{the folder where package binary tar ball stores}
}
\description{
upload the R package binary build as an asset to its github release
}

