% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_proto}
\alias{plot_proto}
\title{pretty layout of proto objects hierarchy}
\usage{
plot_proto(e)
}
\arguments{
\item{e}{environment or proto object}
}
\description{
default layout does not play well with the bigger tree
}

