% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{unlockNamespace}
\alias{unlockNamespace}
\title{unlock a package namespace}
\usage{
unlockNamespace(env)
}
\arguments{
\item{env}{the package namespace to unlock}
}
\description{
unlock a package namespace
}

