% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{updateFunc}
\alias{updateFunc}
\title{update a function within a package namespace}
\usage{
updateFunc(name, pkgName)
}
\arguments{
\item{name}{\code{character} the function name defined in the global session}

\item{pkgName}{\code{character} the package name that the patched function will be writing to}
}
\description{
This is useful for testing a function patch without reinstalling the package or loading entire package into global session
The function has to pre-exist within the package namespace.
}
\examples{
\dontrun{
updateFunc("read.ncdfFlowSet", "ncdfFlow")
}
}

