% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/github_release.R
\name{github_release}
\alias{github_release}
\title{upload the R package binary build as an asset to its github release}
\usage{
github_release(repo, username, token, release_tag = "linux", path = ".",
  ...)
}
\arguments{
\item{repo}{Repository name}

\item{username}{User name}

\item{token}{oAuth2.0 token}

\item{release_tag}{the tag where the asset is associated to}

\item{path}{the folder where package binary tar ball stores}

\item{...}{other arguments passed to httr APIs (i.e. verbose())}
}
\description{
upload the R package binary build as an asset to its github release
}

